build: Modify webpack to split bundles
- Split bundles into main, runtime, and vendor
- Add bundle analyzer to reduce bundle size
- Add index template and dynamic bundles to html
