+++ client/package.json
@@ -16,7 +16,6 @@
   },
   "homepage": "https://github.com/Megaloden106/personal-finance-manager#readme",
   "dependencies": {
-    "@types/d3": "^5.7.2",
     "d3-array": "^2.2.0",
     "d3-ease": "^1.0.5",
     "d3-interpolate": "^1.3.2",
