+++ package.json
@@ -32,6 +32,7 @@
   "dependencies": {},
   "devDependencies": {
     "@types/compression": "0.0.36",
+    "@types/d3": "^5.7.2",
     "@types/express": "^4.17.0",
     "@types/morgan": "^1.7.35",
     "@types/node": "^11.13.17",
